requirements.txt
PyYAML>=6.0.1
atlassian-python-api>=3.41.4
paramiko>=3.3.1
pandas>=2.1.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
html5lib>=1.1
